fix: thread workflow context through live child executor bridge
Propagate the bounded workflow run context into ChildExecutor.Execute so
live provider Infer calls honor session cancellation and maxRunDurationMs,
with unit and durable-session timeout fixtures plus backend-size cleanup.
