+++ pkg/factorysessionexecution/fixtures/runtime_live_child_dispatch_test.go
@@ -43,107 +43,7 @@ func TestJavaScriptRuntimeService_AgentRunLiveChild_ProjectsRealDispatchInspecti
 		t.Fatalf("StartSync: %v", err)
 	}
 
-	read, err := service.GetSession(context.Background(), completed.SessionID)
-	if err != nil {
-		t.Fatalf("GetSession: %v", err)
-	}
-	if read.Progress == nil || read.Progress.TotalDispatches != 1 || read.Progress.CompletedDispatches != 1 {
-		t.Fatalf("progress = %#v, want one completed dispatch", read.Progress)
-	}
-
-	dispatches, err := service.ListDispatches(context.Background(), completed.SessionID)
-	if err != nil {
-		t.Fatalf("ListDispatches: %v", err)
-	}
-	if len(dispatches.Dispatches) != 1 {
-		t.Fatalf("dispatches = %#v, want one dispatch", dispatches.Dispatches)
-	}
-	dispatch := dispatches.Dispatches[0]
-	if dispatch.ID != "dispatch-1" {
-		t.Fatalf("dispatch id = %q, want dispatch-1", dispatch.ID)
-	}
-	if dispatch.Status != fse.DispatchStatusCompleted {
-		t.Fatalf("dispatch status = %q, want COMPLETED", dispatch.Status)
-	}
-	if dispatch.Provider != "mock" {
-		t.Fatalf("dispatch provider = %q, want mock", dispatch.Provider)
-	}
-	if len(dispatch.ProviderSessionRefs) != 1 || dispatch.ProviderSessionRefs[0].ID != "live-provider-session-1" {
-		t.Fatalf("providerSessionRefs = %#v", dispatch.ProviderSessionRefs)
-	}
-
-	dispatchDetail, err := service.GetDispatch(context.Background(), completed.SessionID, dispatch.ID)
-	if err != nil {
-		t.Fatalf("GetDispatch: %v", err)
-	}
-	if dispatchDetail.JavaScript == nil {
-		t.Fatalf("dispatch javascript projection = nil, want execution mode")
-	}
-	if dispatchDetail.JavaScript.ExecutionMode != workflowruntime.ChildExecutionModeLive {
-		t.Fatalf("executionMode = %q, want %q", dispatchDetail.JavaScript.ExecutionMode, workflowruntime.ChildExecutionModeLive)
-	}
-
-	result, err := service.GetResult(context.Background(), completed.SessionID, fse.ResultRequest{
-		Mode: fse.ResultModeFinal,
-	})
-	if err != nil {
-		t.Fatalf("GetResult: %v", err)
-	}
-	primary := decodePrimaryResultMap(t, result.PrimaryResult)
-	child, ok := primary["child"].(map[string]any)
-	if !ok {
-		t.Fatalf("primary child = %#v, want object", primary["child"])
-	}
-	if child["executionMode"] != workflowruntime.ChildExecutionModeLive {
-		t.Fatalf("child executionMode = %#v, want %q", child["executionMode"], workflowruntime.ChildExecutionModeLive)
-	}
-	if child["dispatchId"] != "dispatch-1" {
-		t.Fatalf("child dispatchId = %#v, want dispatch-1", child["dispatchId"])
-	}
-
-	if provider.callCount != 1 {
-		t.Fatalf("provider call count = %d, want 1", provider.callCount)
-	}
-
-	if len(dispatch.OutputArtifactIDs) != 1 || dispatch.OutputArtifactIDs[0] != "child-artifact-1" {
-		t.Fatalf("outputArtifactIds = %#v, want [child-artifact-1]", dispatch.OutputArtifactIDs)
-	}
-	if len(dispatchDetail.StatusTransitions) != 3 {
-		t.Fatalf("statusTransitions = %#v, want queued/running/completed", dispatchDetail.StatusTransitions)
-	}
-	wantTransitions := []fse.DispatchStatus{
-		fse.DispatchStatusQueued,
-		fse.DispatchStatusRunning,
-		fse.DispatchStatusCompleted,
-	}
-	for index, got := range dispatchDetail.StatusTransitions {
-		if got != wantTransitions[index] {
-			t.Fatalf("statusTransitions[%d] = %q, want %q", index, got, wantTransitions[index])
-		}
-	}
-	if len(dispatchDetail.ArtifactIDs) != 1 || dispatchDetail.ArtifactIDs[0] != "child-artifact-1" {
-		t.Fatalf("dispatch artifactIds = %#v, want [child-artifact-1]", dispatchDetail.ArtifactIDs)
-	}
-
-	artifacts, err := service.ListArtifacts(context.Background(), completed.SessionID)
-	if err != nil {
-		t.Fatalf("ListArtifacts: %v", err)
-	}
-	if len(artifacts.Artifacts) != 1 {
-		t.Fatalf("artifacts = %#v, want one child artifact", artifacts.Artifacts)
-	}
-	childArtifact := artifacts.Artifacts[0]
-	if childArtifact.ID != "child-artifact-1" || childArtifact.DispatchID != "dispatch-1" {
-		t.Fatalf("child artifact = %#v", childArtifact)
-	}
-	wantHref := "/factory-sessions/" + completed.SessionID + "/artifacts/child-artifact-1"
-	if childArtifact.RetrievalRef == nil || childArtifact.RetrievalRef.Href != wantHref {
-		t.Fatalf("child artifact retrieval = %#v, want %q", childArtifact.RetrievalRef, wantHref)
-	}
-
-	if read.ArtifactCount != 1 || len(read.ArtifactRefs) != 1 || read.ArtifactRefs[0].ID != "child-artifact-1" {
-		t.Fatalf("session artifact refs = count %d refs %#v", read.ArtifactCount, read.ArtifactRefs)
-	}
+	assertLiveChildDispatchInspection(t, service, completed, provider.callCount)
 }
 
 type fixtureMockProvider struct {
